Organize providers
This separates providers into their own directories to make it easier to find and maintain them.
