+++ src/providers/tfc/authenticationProvider.ts
@@ -7,10 +7,10 @@ import * as os from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import TelemetryReporter from '@vscode/extension-telemetry';
-import { earlyApiClient, TerraformCloudHost, TerraformCloudWebUrl } from '../api/terraformCloud';
+import { earlyApiClient, TerraformCloudHost, TerraformCloudWebUrl } from '../../api/terraformCloud';
 import { isErrorFromAlias, ZodiosError } from '@zodios/core';
-import { apiErrorsToString } from '../api/terraformCloud/errors';
-import { handleZodiosError } from './tfc/uiHelpers';
+import { apiErrorsToString } from '../../api/terraformCloud/errors';
+import { handleZodiosError } from './uiHelpers';
 
 class TerraformCloudSession implements vscode.AuthenticationSession {
   // This id isn't used for anything yet, so we set it to a constant
